+++ eShop.Catalog/Models/Products/CreateProductPrice.cs
@@ -3,7 +3,7 @@
     public class CreateProductPrice
     {
         public Guid CurrencyId { get; set; }
-        public double Price { get; set; }
-        public double? DiscountedPrice { get; set; }
+        public decimal Price { get; set; }
+        public decimal? DiscountedPrice { get; set; }
     }
 }
