+++ eShop.Telegram/TelegramFramework/Controllers/ComissionSettingsController.cs
@@ -74,7 +74,7 @@ namespace eShop.Telegram.TelegramFramework.Controllers
             var user = await _telegramService.GetUserByExternalIdAsync(context.FromId);
             if (user!.AccountId != null)
             {
-                if (double.TryParse(context.Text.Replace(',', '.'), out var amount))
+                if (decimal.TryParse(context.Text.Replace(',', '.'), out var amount))
                 {
                     var request = new SetComissionAmountRequest(user.AccountId.Value, amount);
                     var response = await _setComissionAmounRequestClient.GetResponse<SetComissionAmountResponse>(request);
